refactor: factor out token creation to new user service
This creates a consistent 1:1 handler-to-service mapping as in other parts of the appliation. The Telegram link operation is considered (by design) a user operation, not the app authentication
